end of section 5: added Identity Server

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -42,12 +43,13 @@ public class AuctionController(AuctionDbContext context, IMapper mapper, IPublis
 		return auction is null ? NotFound() : mapper.Map<AuctionDto>(auction);
 	}
 	
+	[Authorize]
 	[HttpPost]
 	public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
 	{
 		var auction = mapper.Map<Auction>(auctionDto);
-		// TODO: add current user as seller
-		auction.Seller = "test-seller";
+		
+		auction.Seller = User.Identity.Name;
 		
 		await context.Auctions.AddAsync(auction);
 		
@@ -62,6 +64,7 @@ public class AuctionController(AuctionDbContext context, IMapper mapper, IPublis
 		return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
 	}
 	
+	[Authorize]
 	[HttpPut("{id}")]
 	public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
 	{
@@ -70,7 +73,9 @@ public class AuctionController(AuctionDbContext context, IMapper mapper, IPublis
 		if (auction is null)
 			return NotFound();
 			
-		// TODO: check seller == username
+		if (auction.Seller != User.Identity.Name)
+			return Forbid();
+			
 		auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
 		auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
 		auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -86,6 +91,7 @@ public class AuctionController(AuctionDbContext context, IMapper mapper, IPublis
 		return res ? Ok() : BadRequest("Failed to save item changes");
 	}
 	
+	[Authorize]
 	[HttpDelete("{id}")]
 	public async Task<ActionResult> DeleteAuction(Guid id)
 	{
@@ -94,7 +100,8 @@ public class AuctionController(AuctionDbContext context, IMapper mapper, IPublis
 		if (auction is null)
 			return NotFound();
 			
-		// TODO: check seller == username
+		if (auction.Seller != User.Identity.Name)
+			return Forbid();
 		
 		context.Auctions.Remove(auction);
